docs: example storybook을 위한 css 추가
